# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileIn": {
        "properties": {
          "connectionId": {
            "format": "uuid",
            "title": "Connectionid",
            "type": "string"
          },
          "effort": {
            "default": "medium",
            "enum": [
              "low",
              "medium",
              "high",
              "max"
            ],
            "title": "Effort",
            "type": "string"
          },
          "fallbackProfileId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackprofileid"
          },
          "maxOutputTokens": {
            "default": 4096,
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Maxoutputtokens",
            "type": "integer"
          },
          "maxRetries": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Maxretries",
            "type": "integer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "offeringId": {
            "format": "uuid",
            "title": "Offeringid",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeoutSeconds": {
            "default": 120,
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Timeoutseconds",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "connectionId",
          "offeringId"
        ],
        "title": "ProfileIn",
        "type": "object"
      },
      "ProfileOut": {
        "properties": {
          "connectionId": {
            "format": "uuid",
            "title": "Connectionid",
            "type": "string"
          },
          "effort": {
            "enum": [
              "low",
              "medium",
              "high",
              "max"
            ],
            "title": "Effort",
            "type": "string"
          },
          "fallbackProfileId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackprofileid"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "maxOutputTokens": {
            "title": "Maxoutputtokens",
            "type": "integer"
          },
          "maxRetries": {
            "title": "Maxretries",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "offeringId": {
            "format": "uuid",
            "title": "Offeringid",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeoutSeconds": {
            "title": "Timeoutseconds",
            "type": "integer"
          },
          "upstreamProvider": {
            "title": "Upstreamprovider",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "connectionId",
          "offeringId",
          "model",
          "upstreamProvider",
          "effort",
          "maxOutputTokens",
          "temperature",
          "timeoutSeconds",
          "maxRetries",
          "fallbackProfileId"
        ],
        "title": "ProfileOut",
        "type": "object"
      },
      "ProfileTestOut": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "costUsd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Costusd",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "inputTokens": {
            "title": "Inputtokens",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "outputTokens": {
            "title": "Outputtokens",
            "type": "integer"
          },
          "providerCostUsd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providercostusd"
          },
          "wasFallback": {
            "title": "Wasfallback",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "content",
          "inputTokens",
          "outputTokens",
          "costUsd",
          "providerCostUsd",
          "wasFallback",
          "errorCode"
        ],
        "title": "ProfileTestOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NexTI Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/profiles": {
      "get": {
        "operationId": "list_profiles_api_v1_ai_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileOut"
                  },
                  "title": "Response List Profiles Api V1 Ai Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "ai"
        ]
      },
      "post": {
        "operationId": "create_profile_api_v1_ai_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/profiles/{profile_id}": {
      "delete": {
        "description": "Assignments that use the profile go with it; profiles that fall back to it lose their fallback.",
        "operationId": "delete_profile_api_v1_ai_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "update_profile_api_v1_ai_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/profiles/{profile_id}:test": {
      "post": {
        "description": "A minimal real call through the gateway; it is recorded in the usage ledger like any other call.",
        "operationId": "test_profile_api_v1_ai_profiles__profile_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Profile",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}